Add ability to save and load documents.
Documents can now be saved and loaded if the program is started with a
-file flag, at which time the user is prompted to enter the name of a
file to use for saving and loading. Currently documents are only loaded
on initialization, but saving happens whenever the user presses Ctrl-s.

Future:
- Load at any time. Loading should update the document of everyone in
the session using document synchronization.
- More integrated file picking. Users should be able to change the file
they want to save to and load from.