Add keybinding for loading at any time.
Ctrl-l now loads the contents of the file as a CRDT document at any
time. To make the implementation cleaner, all passing of the document
between functions is removed. Every modification to the document only
happens to the global document variable, so that an update to the
document in one function easily propogates everywhere.
Right now, one client loading a document doesn't sync up with any other
client in the session, so this implementation doesn't yet work in a
multi-client setting.