feat: export and import document content (#22)
* Add ability to save and load documents.

Documents can now be saved and loaded if the program is started with a
-file flag, at which time the user is prompted to enter the name of a
file to use for saving and loading. Currently documents are only loaded
on initialization, but saving happens whenever the user presses Ctrl-s.

Future:
- Load at any time. Loading should update the document of everyone in
the session using document synchronization.
- More integrated file picking. Users should be able to change the file
they want to save to and load from.

* Add keybinding for loading at any time.

Ctrl-l now loads the contents of the file as a CRDT document at any
time. To make the implementation cleaner, all passing of the document
between functions is removed. Every modification to the document only
happens to the global document variable, so that an update to the
document in one function easily propogates everywhere.
Right now, one client loading a document doesn't sync up with any other
client in the session, so this implementation doesn't yet work in a
multi-client setting.

* Add syncing across all clients on document load.

Whenever a client loads a document with Ctrl-l, the document state is
now synced across all clients. I used the already existing document
syncing channel for this but changed the name to better reflect its
expanded use. Also added handling for the case where there is no file
entered by the client.

* Fix use of wrong error function.

* Add error handling to load and save functions.

* Add unit test for loading CRDT doc from file.

* fix: change file permissions for saving content

* fix: use mutexes for incrementing site ID

Co-authored-by: Ben Muthalaly <benmuthalaly@gmail.com>
Co-authored-by: burntcarrot <aadhav.n1@gmail.com>