Improve error handling for server.
Fixes a problem where the server would be left in a bad state after some
server-side closures of client WebSocket connections. The bad state was
caused by clients not being removed from the activeClients map and
causing nil pointer errors when attempting to broadcast a message, for
example.

This commit tries to avoid that by defining a function that cleanly
disconnects from a client connection. It also refactors some error
handling and logging.