Fix potential concurrency bug with mutex.
I think there was a potential concurrency bug that the mutex previously
wouldn't have addressed: If two clients joined at the same time, the
mutex would avoid a data race on the value of the  site ID, but it was
still possible for the client to be assigned the wrong site ID value
after each mutex had released its lock. I moved the mutex unlock until
after the site ID is recorded in the activeClients map to try and avoid
this issue.

I also moved and edited a log message to make the server logs make a bit
more sense.