Fix cleanup of server state on client disconnect (#27)
* Improve error handling for server.

Fixes a problem where the server would be left in a bad state after some
server-side closures of client WebSocket connections. The bad state was
caused by clients not being removed from the activeClients map and
causing nil pointer errors when attempting to broadcast a message, for
example.

This commit tries to avoid that by defining a function that cleanly
disconnects from a client connection. It also refactors some error
handling and logging.

* Fix potential concurrency bug with mutex.

I think there was a potential concurrency bug that the mutex previously
wouldn't have addressed: If two clients joined at the same time, the
mutex would avoid a data race on the value of the  site ID, but it was
still possible for the client to be assigned the wrong site ID value
after each mutex had released its lock. I moved the mutex unlock until
after the site ID is recorded in the activeClients map to try and avoid
this issue.

I also moved and edited a log message to make the server logs make a bit
more sense.

* Add missing return after write-to-server error.