Add updating local cursor on remote insert/delete
Cursor now "stays in place" when another user adds or deletes
characters.

This needs to be tested when two people are editing simultaneously to
make sure behavior is as desired.

Also, I wasn't sure what should happen if two users have their cursor on
the same character and one user does an insert. Should the other user's
cursor stay in place or move along with the inserts? Currently, it moves
along with the inserts.