Add basic scrolling.
Added basic scrolling with the cursor. When the cursor reaches the
border of the editor window, the editor window will shift to follow the
cursor. This level of basic scrolling seems to be on par with the kilo
text editor, as a point of comparison.

Also added some basic test cases. More tests would be welcome.

More robust scrolling in the future could include the ability to scroll
independently of the cursor, or scroll multiple lines at once.