Cursor updating and scrolling (#28)
* Add updating local cursor on remote insert/delete

Cursor now "stays in place" when another user adds or deletes
characters.

This needs to be tested when two people are editing simultaneously to
make sure behavior is as desired.

Also, I wasn't sure what should happen if two users have their cursor on
the same character and one user does an insert. Should the other user's
cursor stay in place or move along with the inserts? Currently, it moves
along with the inserts.

* (WIP) Implement scrolling.

This commit is to store the progress of an attempt to add scrolling as a
full feature, including scrolling independently of the cursor.

I've decided to scale back the scope and only try to implement scrolling
with the cursor for now, which is how the kilo editor does it.

* Add basic scrolling.

Added basic scrolling with the cursor. When the cursor reaches the
border of the editor window, the editor window will shift to follow the
cursor. This level of basic scrolling seems to be on par with the kilo
text editor, as a point of comparison.

Also added some basic test cases. More tests would be welcome.

More robust scrolling in the future could include the ability to scroll
independently of the cursor, or scroll multiple lines at once.

* Remove logging, clean up extra functions/comments.

* Fix formatting issue.

* Add 'jumping' scroll.

Window now follows cursor even if it has to 'jump' more than one
row/column at a time.

* Make scrolling toggleable with a flag.

* refactor: use a config struct

* fix: set ScrollEnabled to true in test

---------

Co-authored-by: burntcarrot <aadhav.n1@gmail.com>