Remove unused method and add mutex. (#30)
* Remove unused AddRune method.

The AddRune method was used for local insertions to update the
editor's slice of runes, bypassing the CRDT doc entirely. The
editor's slice of runes was then being immediately overwritten by
a CRDT insert, so AddRune was doing nothing.

This removes the AddRune method and slightly adapts the local
insertion logic to keep behavior the same.

* Add mutex around local clock increment.

The GenerateInsert method increments the local clock. Without a  mutex,
a race condition was possible because both the local and message
handling threads were able to insert characters simultaneously.